requirements.txt
Django==6.0.1
playwright==1.48.0
python-pptx==1.0.2
